pull out shared deploy code into deploy utils (#34321)
